Add initial work on script to follow conversations.
The overall idea is that this script will run a loop dumping
conversation contents every minute, then process dump to add
new messages to log file. This way, the user won't lose the
contents of a conversation.

Additionally, there should be a way to define hooks to be
executed upon arrival of new messages. For notification
purpospes etc.

This is nowhere near done yet.